+++ CHANGELOG.md
@@ -38,6 +38,9 @@ to function correctly. [PR #1231](https://github.com/3scale/APIcast/pull/1231)
 ## [3.9.1] 2020-10-13
 
 - Added a new metric when the `worker_process` starts [PR #1228](https://github.com/3scale/APIcast/pull/1228) [THREESCALE-5965](https://issues.redhat.com/browse/THREESCALE-5965)
+- Fixed issues when using fully qualified DNS query [PR #1235](https://github.com/3scale/APIcast/pull/1235) [THREESCALE-4752](https://issues.redhat.com/browse/THREESCALE-4752)
+- Fixed issues with OIDC validation [PR #1239](https://github.com/3scale/APIcast/pull/1239) [THREESCALE-6313](https://issues.redhat.com/browse/THREESCALE-6313)
+
 
 
 ## [3.9.0] 2020-08-17